+++ spec/factories/access_grant.rb
@@ -1,5 +1,5 @@
 FactoryGirl.define do
-  factory :access_grant do
+  factory :access_grant, :class => Doorkeeper::AccessGrant do
     sequence(:resource_owner_id) { |n| n }
     application
     redirect_uri "https://app.com/callback"
