+++ spec/factories/access_token.rb
@@ -1,5 +1,5 @@
 FactoryGirl.define do
-  factory :access_token do
+  factory :access_token, :class => Doorkeeper::AccessToken do
     sequence(:resource_owner_id) { |n| n }
     application
     expires_in 2.hours
