+++ spec/factories/application.rb
@@ -1,5 +1,5 @@
 FactoryGirl.define do
-  factory :application do
+  factory :application, :class => Doorkeeper::Application do
     sequence(:name){ |n| "Application #{n}" }
     redirect_uri "https://app.com/callback"
   end
