+++ spec/requests/flows/authorization_code_spec.rb
@@ -16,7 +16,7 @@ feature 'Authorization Code Flow' do
 
     i_should_be_on_client_callback(@client)
 
-    url_should_have_param("code", AccessGrant.first.token)
+    url_should_have_param("code", Doorkeeper::AccessGrant.first.token)
     url_should_not_have_param("state")
     url_should_not_have_param("error")
   end
@@ -24,7 +24,7 @@ feature 'Authorization Code Flow' do
   scenario 'resource owner authorizes the client with state parameter set' do
     visit authorization_endpoint_url(:client => @client, :state => "return-me")
     click_on "Authorize"
-    url_should_have_param("code", AccessGrant.first.token)
+    url_should_have_param("code", Doorkeeper::AccessGrant.first.token)
     url_should_have_param("state", "return-me")
   end
 
@@ -32,43 +32,43 @@ feature 'Authorization Code Flow' do
     client_is_authorized(@client, @resource_owner)
     visit authorization_endpoint_url(:client => @client)
 
-    authorization_code = AccessGrant.first.token
+    authorization_code = Doorkeeper::AccessGrant.first.token
     post token_endpoint_url(:code => authorization_code, :client => @client)
 
-    AccessToken.count.should be(1)
+    Doorkeeper::AccessToken.count.should be(1)
 
-    should_have_json 'access_token', AccessToken.first.token
+    should_have_json 'access_token', Doorkeeper::AccessToken.first.token
   end
 
   scenario 'revokes and return new token if it is has expired' do
     client_is_authorized(@client, @resource_owner)
-    token = AccessToken.first
+    token = Doorkeeper::AccessToken.first
     token.update_attribute :expires_in, -100
     visit authorization_endpoint_url(:client => @client)
 
-    authorization_code = AccessGrant.first.token
+    authorization_code = Doorkeeper::AccessGrant.first.token
     post token_endpoint_url(:code => authorization_code, :client => @client)
 
     token.reload.should be_revoked
-    AccessToken.count.should be(2)
+    Doorkeeper::AccessToken.count.should be(2)
 
-    should_have_json 'access_token', AccessToken.last.token
+    should_have_json 'access_token', Doorkeeper::AccessToken.last.token
   end
 
   scenario 'resource owner requests an access token with authorization code' do
     visit authorization_endpoint_url(:client => @client)
     click_on "Authorize"
 
-    authorization_code = AccessGrant.first.token
+    authorization_code = Doorkeeper::AccessGrant.first.token
     post token_endpoint_url(:code => authorization_code, :client => @client)
 
     access_token_should_exists_for(@client, @resource_owner)
 
     should_not_have_json 'error'
 
-    should_have_json 'access_token', AccessToken.first.token
+    should_have_json 'access_token', Doorkeeper::AccessToken.first.token
     should_have_json 'token_type',   "bearer"
-    should_have_json 'expires_in',   AccessToken.first.expires_in
+    should_have_json 'expires_in',   Doorkeeper::AccessToken.first.expires_in
 
     should_not_have_json 'refresh_token'
   end
@@ -96,7 +96,7 @@ feature 'Authorization Code Flow' do
       visit authorization_endpoint_url(:client => @client, :scope => "public write")
       click_on "Authorize"
 
-      authorization_code = AccessGrant.first.token
+      authorization_code = Doorkeeper::AccessGrant.first.token
       post token_endpoint_url(:code => authorization_code, :client => @client)
 
       access_token_should_exists_for(@client, @resource_owner)
@@ -108,12 +108,12 @@ feature 'Authorization Code Flow' do
       visit authorization_endpoint_url(:client => @client, :scope => "public")
       click_on "Authorize"
 
-      authorization_code = AccessGrant.first.token
+      authorization_code = Doorkeeper::AccessGrant.first.token
       post token_endpoint_url(:code => authorization_code, :client => @client)
 
-      AccessToken.count.should be(2)
+      Doorkeeper::AccessToken.count.should be(2)
 
-      should_have_json 'access_token', AccessToken.last.token
+      should_have_json 'access_token', Doorkeeper::AccessToken.last.token
     end
   end
 end
